hoan thien product

+++ src/main/java/com/casestudy5/model/entity/image/ImageDTO.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class ImageDTO {
-    private Integer id; // Thêm trường id
+    private long id; // Thêm trường id
     private String imageType;
     private String fileName;
